# PERF: Cython 3.0 regression with frame constructor