- [ ] closes #xxxx (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

Handles the 5 `frame_ctor.FromDicts` regressions in #55179

There is only one use of this function:

https://github.com/pandas-dev/pandas/blob/b2d9ec17c52084ee2b629633c9119c01ea11d387/pandas/core/internals/construction.py#L627

and it's guaranteed to always have object dtype:

https://github.com/pandas-dev/pandas/blob/b2d9ec17c52084ee2b629633c9119c01ea11d387/pandas/core/internals/construction.py#L618-L619

ASVs on this branch:

```
| Change   | Before [b2d9ec17] <perf_cython3_frame_dict_constructor~1>   | After [951c5bfa] <perf_cython3_frame_dict_constructor>   |   Ratio | Benchmark (Parameter)                               |
|----------|-------------------------------------------------------------|----------------------------------------------------------|---------|-----------------------------------------------------|
| -        | 17.2±0.2ms                                                  | 15.3±0.3ms                                               |    0.89 | frame_ctor.FromDicts.time_nested_dict               |
| -        | 13.2±0.2ms                                                  | 11.7±0.3ms                                               |    0.89 | frame_ctor.FromDicts.time_nested_dict_index_columns |
| -        | 13.1±0.2ms                                                  | 11.3±0.3ms                                               |    0.86 | frame_ctor.FromDicts.time_nested_dict_index         |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```